moved add modification code from design_main_strand.dart to design_main_strand_modification.dart to keep similar code together
